handle userid not found condition for api 12

+++ model/interest.js
@@ -21,34 +21,43 @@ let Interest = {
 			if (err) {
 				return callback(err, null);
 			} else {
-				console.log("Connection established!");
-				let catArr = int.split(","); // Returns array of all categoryid as an array.
-				// // Handles duplicate values from request body
-				// let dltQueryStr = "";
-				let dltQueryVal = [uid];
-				const dltQuery = "DELETE FROM interest WHERE fk_user_id = ?;\n";
-				// for (let i = 0; i < catArr.length; i++) {
-				//     dltQueryStr += dltQuery;
-				//     dltQueryVal.push(uid, parseInt(catArr[i],10));
-				// };
-				//----------------------------------------------
-				// Duplicate queries to match request body
-				let queryStr = "";
-				let queryVal = [];
-				const query = `INSERT INTO interest (fk_user_id, fk_category_id) VALUES (?, ?);\n`;
-				for (let i = 0; i < catArr.length; i++) {
-					queryStr += query;
-					queryVal.push(uid, parseInt(catArr[i], 10));
-				}
-				//----------------------------------------------
-				sql = dltQuery + queryStr;
-				queryArray = dltQueryVal.concat(queryVal);
-				conn.query(sql, queryArray, (error, result) => {
-					conn.end();
+				var sql = "SELECT userid FROM user WHERE userid = ?"; 
+				conn.query(sql, [uid], function (error, result) {
 					if (error) {
+						conn.end();
 						return callback(error, null);
-					} else {
-						return callback(null, result);
+					} else { 
+						if(result.length == 0) return callback(null, null); // no such user
+						console.log("Connection established!");
+						let catArr = int.split(","); // Returns array of all categoryid as an array.
+						// // Handles duplicate values from request body
+						// let dltQueryStr = "";
+						let dltQueryVal = [uid];
+						const dltQuery = "DELETE FROM interest WHERE fk_user_id = ?;\n";
+						// for (let i = 0; i < catArr.length; i++) {
+						//     dltQueryStr += dltQuery;
+						//     dltQueryVal.push(uid, parseInt(catArr[i],10));
+						// };
+						//----------------------------------------------
+						// Duplicate queries to match request body
+						let queryStr = "";
+						let queryVal = [];
+						const query = `INSERT INTO interest (fk_user_id, fk_category_id) VALUES (?, ?);\n`;
+						for (let i = 0; i < catArr.length; i++) {
+							queryStr += query;
+							queryVal.push(uid, parseInt(catArr[i], 10));
+						}
+						//----------------------------------------------
+						sql = dltQuery + queryStr;
+						queryArray = dltQueryVal.concat(queryVal);
+						conn.query(sql, queryArray, (error, result) => {
+							conn.end();
+							if (error) {
+								return callback(error, null);
+							} else {
+								return callback(null, result);
+							}
+						});
 					}
 				});
 			}
